feat: integrate Telegram Bot logic

+++ src/modules/cron/cron.service.ts
@@ -5,13 +5,15 @@ import { PrismaService } from '@/src/core/prisma/prisma.service'
 
 import { MailService } from '../libs/mail/mail.service'
 import { StorageService } from '../libs/storage/storage.service'
+import { TelegramService } from '../libs/telegram/telegram.service'
 
 @Injectable()
 export class CronService {
 	public constructor(
 		private readonly prismaService: PrismaService,
 		private readonly mailService: MailService,
-		private readonly storageService: StorageService
+		private readonly storageService: StorageService,
+		private readonly telegramService: TelegramService
 	) {}
 
 	// @Cron('*/10 * * * * *') // Every 10 secs
@@ -27,12 +29,31 @@ export class CronService {
 				deactivatedAt: {
 					lte: sevenDaysAgo
 				}
+			},
+			include: {
+				notificationSettings: true,
+				streams: true
 			}
 		})
 		for (const user of deactivatedAccounts) {
 			// await this.mailService.sendAccountDeletionMail(user.email) /** COMMENTED TO AVOID EMAIL VERIFICATION */
 
-			await this.storageService.remove(user.avatar)
+			if (
+				user.notificationSettings.telegramNotifications &&
+				user.telegramId
+			) {
+				await this.telegramService.sendAccountDeletedMessage(
+					user.telegramId
+				)
+			}
+
+			if (user.avatar) {
+				await this.storageService.remove(user.avatar)
+			}
+
+			if (user.streams.thumbnailUrl) {
+				await this.storageService.remove(user.streams.thumbnailUrl)
+			}
 		}
 
 		await this.prismaService.user.deleteMany({
